requirements.txt
telethon==1.21
jinja2==2.11.3
PyYAML==5.4.1
cryptg==0.2.post2
Pillow==8.2.0
feedgen==0.9.0
pandas==1.2.4
openpyxl==3.0.7
pytz
pymongo